% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishingLines.R
\name{fishingLines}
\alias{fishingLines}
\title{Redraw trade lines}
\usage{
fishingLines(spLines)
}
\arguments{
\item{spLines}{Vector of SpatialLine/gcIntermediate functions populated in fishTrade()}
}
\value{
An object of class 'SpatialLinesDataFrame' to be appended to leaflet
}
\description{
Redraw trade lines
}
\examples{
someExample <- 'goes here' 
}

