% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishReadR.R
\name{fishReadR}
\alias{fishReadR}
\title{Read txt data}
\usage{
fishReadR(userData)
}
\arguments{
\item{userData}{A data.frame or a string path to a file (eventually an R data type instead/in addition, but for now we use local text file)}
}
\value{
TBD
}
\description{
Read txt data
}
\examples{
someExample <- 'goes here' 
}

