% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishRmap.R
\name{fishRmap}
\alias{fishRmap}
\title{Call visualization function}
\usage{
fishRmap(userdata, import = "Import", export = "Export",
  species = "Species", value = "value", year = "Year", iso = "numeric",
  importOnly = FALSE, exportOnly = FALSE)
}
\arguments{
\item{userdata}{A string path to the file containing fishery data}

\item{import}{Character string name of column with three-character or -digit ISO code of country importing fish}

\item{export}{Character string name of column with three-character or -digit ISO code of country exporting fish}

\item{species}{Character string name of column with fish species names}

\item{value}{Character string name of column with value of trade}

\item{year}{Character string name of column with year}

\item{iso}{Type of ISO country identifier used in "import" and "export" columns - either "character" or "numeric"}
}
\value{
TBD
}
\description{
Call visualization function
}
\examples{
someExample <- 'goes here' 
}
