% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishUp.r
\name{fishUp}
\alias{fishUp}
\title{Call update function to create hard links to package files in shiny shared dir}
\usage{
fishUp(fishrDir, sroot)
}
\arguments{
\item{sroot}{A string path to the shiny shared directory}

\item{fishDir}{A string path to the directory of the fishRmap library}
}
\value{
An empty string
}
\description{
Call update function to create hard links to package files in shiny shared dir
}
\examples{
someExample <- 'goes here' 
}
